pyproject.toml
[tool.poetry]
name = "pycarla"
version = "0.3.1"
description = "Use VST/LV2/etc. plugins with realtime abilities in Python"
authors = ["Federico Simonetta <federico.simonetta@unimi.it>"]
license = "GPL-3.0-only"
readme = "README.rst"
homepage = "https://pycarla.readthedocs.org"
documentation = "https://pycarla.readthedocs.org"
repository = "https://github.com/00sapo/pycarla"
classifiers = [
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Multimedia :: Sound/Audio :: MIDI",
    "Intended Audience :: Science/Research",
    "Topic :: Multimedia :: Sound/Audio :: Sound Synthesis",
    "Topic :: Multimedia :: Sound/Audio :: Capture/Recording",
    "Topic :: Multimedia :: Sound/Audio :: Players",
    "Topic :: Multimedia :: Sound/Audio :: Mixers",
    "Intended Audience :: Developers",
    "Intended Audience :: Other Audience"
]

[tool.poetry.dependencies]
python = "^3.6"
jack-client = "^0.5.2"
numpy = "^1.19.0"
pysoundfile = "^0.9.0"
PyQt5 = "^5.15.2"
mido = "^1.2.9"
psutil = "^5.8.0"

[tool.poetry.dev-dependencies]
ipdb = "^0.13.2"
ipython = "^7.15.0"
